gulpfile moved outside. Various changes to make `gulp dev` prepare a dev environment. For now, it only starts docker-compose. Soon, it'll watch files on the host (which will be reflected to the containers given they use volumes for code)
