+++ package.json
@@ -5,15 +5,11 @@
   "main": "index.js",
   "scripts": {
     "postinstall": "cp ../git-hooks/pre-commit ../.git/hooks/pre-commit; chmod a+x ../.git/hooks/pre-commit",
-    "start-admin-dev": "gulp admin-dev --gulpfile core/gulpfile.js",
-    "start-admin-prod": "gulp admin-prod --gulpfile core/gulpfile.js",
-    "start-dashboard-dev": "gulp dashboard-dev --gulpfile core/gulpfile.js",
-    "start-dashboard-prod": "gulp dashboard-prod --gulpfile core/gulpfile.js",
-    "start-endpoint-dev": "gulp serve-endpoint --gulpfile core/gulpfile.js",
-    "start-endpoint-prod": "gulp serve-endpoint --gulpfile core/gulpfile.js",
-    "start-citizen-dev": "gulp citizen-dev --gulpfile core/gulpfile.js",
-    "start-citizen-prod": "gulp citizen-prod --gulpfile core/gulpfile.js",
-    "init-db": "gulp init --gulpfile core/gulpfile.js",
+    "start-admin": "node core/server/admin.js",
+    "start-dashboard": "node core/server/dashboard.js",
+    "start-endpoint": "node core/server/endpoint.js",
+    "start-citizen": "node core/server/citizen.js",
+    "init-db": "node tools/init-db.js",
     "lint": "eslint ./",
     "test": "mocha test/automated/*",
     "validate": "npm ls"
@@ -46,7 +42,7 @@
     "es6-shim": "^0.31.2",
     "express": "^4.12.4",
     "flux": "^2.0.3",
-    "gulp": "~3.8.8",
+    "gulp": "~3.9.0",
     "gulp-express": "git://github.com/gimm/gulp-express.git#9a3c842b55246fe039c85b56b4a8e5401ce779bf",
     "gulp-livereload": "~2.1.1",
     "gulp-plumber": "^1.0.1",
